CANoe project foundation
Setup a CANdb which contains mileage single along with Mileage generator ECU and ICP ECU.

The Mileage generator is the transmitter node for the mileage signal and the ICP is the receiver node.

I wrote some simple capl code for each node that is used to generate mileage and send it onto the bus. The receiver node then receives the mileage each time its transmitted and which is the ICP. The ICP then keeps adding the mileage up and Writes out the total mileage each time it receives mileage.